- Updated Renderer interface to use Floating point values for position information. Closes #72.

+++ core/include/NAS2D/Resources/Sprite.h
@@ -50,7 +50,7 @@ public:
 	void pause();
 	void resume();
 
-	void update(int x, int y);
+	void update(float x, float y);
 
 	void rotation(float angle);
 	float rotation() const;
